Add GuestAddressSpaceView concept
Emulated devices should not need to access arbitrary vcpu state,
but _do_ need to be able to translate guest addresses. This
requires that they at least have access to cr3/cr0 (among other
things). The GuestAddressSpaceView is a view of the GuestAddressSpace
that is effectively bound to some particular vcpu state. This
allows a device to interact with guest memory with having
access to arbitrary vcpu state.