Add basic support for ticking time in guest context
This requires an extensive set of changes. We must add support
for a timer wheel that is serviced by the local apic timer. This
requires support for per-core variables so we have a practical way
of accessing the lapic and wheel from a variety of locations.

We must also add a much more functional PIT implementation, so the
guest can setup one-shot or periodic timers. However this change does
_not_ add additional PIC support. Therefore, we just hard code the
linux timer interrupt vector, as it will not be remapped in practice.

The current intended functionality is that the hypervisor will mask
interrupts and always service interrupts by handling the vmexit
caused by any external interrupt. However, this may not work in
practice, so this change adds better interrupt handling support.