Prevent access to per-core variables without proper macro
Without this patch, it's possible to declare a per-core variable
like so:

    declare_per_core! {
        static mut TIMER_WHEEL: Option<TimerWheel> = None;
    }

Then access the variable like:

    unsafe {
        TIMER_WHEEL = Some(value);
    }

But this will always mutate the value on the BSP core (and is not
thread-safe). This patch generates a new symbol for the actual
static variable, which effectively requires the usage of the
'get_per_core' or 'get_per_core_mut' and prevents this category
of error.